Better code
Clean now.
