+++ src/accumulator.h
