+++ src/main.cpp
@@ -1,0 +1,6 @@
+#include "uci.h"
+
+int main(int argc, char* argv[]) {
+    uci_loop();
+    return 0;
+}
